[ITG-144] Eliminate cyclic imports between models (#45)
* Refactor models to eliminate cyclic imports

* Update resources code

* Remove redundant import
