+++ bode/bode/models/task_relation/actions.py
@@ -1,14 +1,48 @@
 from sqlalchemy import or_
 
 from bode.app import db
-from bode.models.task import TaskStatus
-from bode.models.task_relation import DirectedRelationType, RelationType, TaskRelation
+from bode.models.enums import SYMMETRIC_RELATIONS, RelationType, TaskStatus
+from bode.models.task.model import Task
+from bode.models.task_relation.model import TaskRelation
 from bode.utils.graph import Graph
 
 
-def get_lhs_related_tasks(task_id, filters=list()):
-    from bode.models.task import Task
+def create_task_relation(**relation_data):
+    relation = TaskRelation(**relation_data)
 
+    db.session.add(relation)
+    if relation_data["type"] in SYMMETRIC_RELATIONS:
+        symmetric_data = relation_data.copy()
+        symmetric_data["first_task_id"] = relation_data["second_task_id"]
+        symmetric_data["second_task_id"] = relation_data["first_task_id"]
+
+        symmetric_relation = TaskRelation(**symmetric_data)
+        db.session.add(symmetric_relation)
+
+    db.session.commit()
+
+    return relation
+
+
+def delete_task_relation(relation_id):
+    relation = TaskRelation.query.get_or_404(relation_id)
+
+    if relation.type in SYMMETRIC_RELATIONS:
+        symmetric_relation = TaskRelation.query.filter(
+            TaskRelation.first_task_id == relation.second_task_id,
+            TaskRelation.second_task_id == relation.first_task_id,
+            TaskRelation.type == relation.type,
+        ).one()
+
+        db.session.delete(symmetric_relation)
+
+    db.session.delete(relation)
+    db.session.commit()
+
+    return relation
+
+
+def get_lhs_related_tasks(task_id, filters=list()):
     all_filters = [TaskRelation.first_task_id == task_id] + filters
     return (
         db.session.query(TaskRelation, Task)
@@ -19,8 +53,6 @@ def get_lhs_related_tasks(task_id, filters=list()):
 
 
 def get_rhs_related_tasks(task_id, filters=list()):
-    from bode.models.task import Task
-
     all_filters = [TaskRelation.second_task_id == task_id] + filters
     return (
         db.session.query(TaskRelation, Task)
@@ -34,29 +66,14 @@ def get_related_tasks(task_id):
     return get_lhs_related_tasks(task_id) + get_rhs_related_tasks(task_id)
 
 
-def map_to_related_task_schema(relation: TaskRelation, task_id):
-    match relation.type:
-        case RelationType.Dependent.value:
-            if task_id == relation.first_task_id:
-                return DirectedRelationType.IsBlockedBy.value
-            return DirectedRelationType.Blocks.value
-        case RelationType.Subtask.value:
-            if task_id == relation.first_task_id:
-                return DirectedRelationType.Supertask.value
-            return DirectedRelationType.Subtask.value
-        case _:
-            return DirectedRelationType.Interchangable.value
-
-
-def get_relation_types(task_id):
-    relations_lhs = db.session.query(TaskRelation).filter(TaskRelation.first_task_id == task_id).all()
-    relations_rhs = db.session.query(TaskRelation).filter(TaskRelation.second_task_id == task_id).all()
-    return {map_to_related_task_schema(relation, task_id) for relation in relations_lhs + relations_rhs}
+def is_task_blocked(task_id):
+    for relation, task in get_lhs_related_tasks(task_id):
+        if relation.type == RelationType.Dependent.value and task.status == TaskStatus.TODO.value:
+            return True
+    return False
 
 
 def get_transitive_interchangable_related_tasks(task_id, filters=list()):
-    from bode.models.task import Task
-
     def get_distinct_uuids(uuids):
         return list(set(uuids))
 
@@ -97,10 +114,3 @@ def get_transitive_interchangable_related_tasks(task_id, filters=list()):
         .distinct(Task.id)
         .all()
     )
-
-
-def is_task_blocked(task_id):
-    for relation, task in get_lhs_related_tasks(task_id):
-        if relation.type == RelationType.Dependent.value and task.status == TaskStatus.TODO.value:
-            return True
-    return False
