+++ bode/bode/resources/tags/api.py
@@ -2,7 +2,8 @@ from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import IntegrityError
 
-from bode.models.tag import Tag
+from bode.models.tag.actions import create_tag, delete_tag
+from bode.models.tag.model import Tag
 from bode.resources.tags.schemas import TagInputSchema, TagSchema
 
 blueprint = Blueprint("tags", "tags", url_prefix="/tags")
@@ -18,7 +19,7 @@ class Tags(MethodView):
     @blueprint.response(201, TagSchema)
     def post(self, tag_data):
         try:
-            return Tag.create(tag_data["name"])
+            return create_tag(tag_data["name"])
         except IntegrityError:
             abort(409, message="Item already exists")
 
@@ -27,4 +28,4 @@ class Tags(MethodView):
 class TagsById(MethodView):
     @blueprint.response(200, TagSchema)
     def delete(self, tag_id):
-        return Tag.delete(tag_id)
+        return delete_tag(tag_id)
